Lab10/requirements.txt
pygame==2.5.2
psycopg2-binary==2.9.9
pandas==2.1.1
